Move vcs/git/commits/CommitGraph to gitserver/commands/CommitGraph
