+++ internal/gitserver/commands.go
@@ -3,10 +3,12 @@ package gitserver
 import (
 	"bytes"
 	"context"
+	"fmt"
 	"io"
 	"net/mail"
 	"strconv"
 	"strings"
+	"time"
 
 	"github.com/sourcegraph/go-diff/diff"
 
@@ -212,6 +214,36 @@ func checkSpecArgSafety(spec string) error {
 		return errors.Errorf("invalid git revision spec %q (begins with '-')", spec)
 	}
 	return nil
+}
+
+// CommitGraph returns the commit graph for the given repository as a mapping
+// from a commit to its parents. If a commit is supplied, the returned graph will
+// be rooted at the given commit. If a non-zero limit is supplied, at most that
+// many commits will be returned.
+func (c *ClientImplementor) CommitGraph(ctx context.Context, repo api.RepoName, opts gitdomain.CommitGraphOptions) (_ *gitdomain.CommitGraph, err error) {
+	args := []string{"log", "--pretty=%H %P", "--topo-order"}
+	if opts.AllRefs {
+		args = append(args, "--all")
+	}
+	if opts.Commit != "" {
+		args = append(args, opts.Commit)
+	}
+	if opts.Since != nil {
+		args = append(args, fmt.Sprintf("--since=%s", opts.Since.Format(time.RFC3339)))
+	}
+	if opts.Limit > 0 {
+		args = append(args, fmt.Sprintf("-%d", opts.Limit))
+	}
+
+	cmd := c.Command("git", args...)
+	cmd.Repo = repo
+
+	out, err := cmd.CombinedOutput(ctx)
+	if err != nil {
+		return nil, err
+	}
+
+	return gitdomain.ParseCommitGraph(strings.Split(string(out), "\n")), nil
 }
 
 // DevNullSHA 4b825dc642cb6eb9a060e54bf8d69288fbee4904 is `git hash-object -t
